optimize NumericRangeGroupCounter::addDoc()
1. reduce frequency of division operation
2. use int instead of int64_t as a price doesn't seems like an overflow

+++ source/core/mining-manager/faceted-submanager/NumericRangeGroupCounter.h
@@ -22,7 +22,7 @@ struct Log10SegmentTree
 {
     Log10SegmentTree();
 
-    void insertPoint(int64_t point);
+    void insertPoint(int point);
 
     int level0_;
     int level1_[LEVEL_1_OF_SEGMENT_TREE];
